+++ Applications/SaveDataApp.cs
@@ -8,37 +8,54 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.Runtime.Serialization;
+using LCU.Graphs.Registry.Enterprises.Apps;
 using LCU.Personas.Client.Applications;
 using Fathym;
 using Microsoft.Azure.WebJobs.Extensions.SignalRService;
 using Microsoft.WindowsAzure.Storage.Blob;
 using LCU.StateAPI.Utilities;
+using LCU.State.API.NapkinIDE.ApplicationManagement.State;
 
-namespace LCU.State.API.NapkinIDE.ApplicationManagement
+namespace LCU.State.API.NapkinIDE.ApplicationManagement.Applications
 {
     [Serializable]
     [DataContract]
-    public class SetViewTypeRequest
+    public class SaveDataAppRequest
     {
         [DataMember]
-        public virtual DAFAppTypes AppType { get; set; }
+        public virtual Application App { get; set; }
     }
 
-    public class SetViewType
+    public class SaveDataApp
     {
-        [FunctionName("SetViewType")]
+        protected ApplicationDeveloperClient appDev;
+
+        protected ApplicationManagerClient appMgr;
+
+        public SaveDataApp(ApplicationDeveloperClient appDev, ApplicationManagerClient appMgr)
+        {
+            this.appDev = appDev;
+            
+            this.appMgr = appMgr;
+        }
+
+        [FunctionName("SaveDataApp")]
         public virtual async Task<Status> Run([HttpTrigger] HttpRequest req, ILogger log,
             [SignalR(HubName = ApplicationManagementState.HUB_NAME)]IAsyncCollector<SignalRMessage> signalRMessages,
             [Blob("state-api/{headers.lcu-ent-api-key}/{headers.lcu-hub-name}/{headers.x-ms-client-principal-id}/{headers.lcu-state-key}", FileAccess.ReadWrite)] CloudBlockBlob stateBlob)
         {
-            return await stateBlob.WithStateHarness<ApplicationManagementState, SetViewTypeRequest, ApplicationManagementStateHarness>(req, signalRMessages, log,
+            return await stateBlob.WithStateHarness<ApplicationManagementState, SaveDataAppRequest, ApplicationManagementStateHarness>(req, signalRMessages, log,
                 async (harness, reqData, actReq) =>
             {
                 var stateDetails = StateUtils.LoadStateDetails(req);
 
-                log.LogInformation($"Setting View Type: {reqData.AppType}");
+                // appMgr.SetAuthorization(stateDetails)
 
-                await harness.SetViewType(reqData.AppType);
+                await harness.SaveDataApp(appDev, appMgr, stateDetails.EnterpriseAPIKey, stateDetails.Host, reqData.App);
+
+                log.LogInformation($"Saving Data App: {reqData.App.Name}");
+
+                await harness.LoadApplications(appMgr, stateDetails.EnterpriseAPIKey);
 
                 return Status.Success;
             });
