+++ Applications/SetViewType.cs
@@ -8,49 +8,38 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.Runtime.Serialization;
-using LCU.Graphs.Registry.Enterprises.Apps;
+using LCU.Personas.Client.Applications;
+using Fathym;
 using Microsoft.Azure.WebJobs.Extensions.SignalRService;
 using Microsoft.WindowsAzure.Storage.Blob;
-using Fathym;
 using LCU.StateAPI.Utilities;
-using LCU.Personas.Client.Applications;
+using LCU.State.API.NapkinIDE.ApplicationManagement.State;
 
-namespace LCU.State.API.NapkinIDE.ApplicationManagement
+namespace LCU.State.API.NapkinIDE.ApplicationManagement.Applications
 {
     [Serializable]
     [DataContract]
-    public class SaveDAFAppRequest
+    public class SetViewTypeRequest
     {
         [DataMember]
-        public virtual DAFApplicationConfiguration DAFApp { get; set; }
+        public virtual DAFAppTypes AppType { get; set; }
     }
 
-    public class SaveDAFApp
+    public class SetViewType
     {
-        protected ApplicationDeveloperClient appDev;
-
-        protected ApplicationManagerClient appMgr;
-
-        public SaveDAFApp(ApplicationDeveloperClient appDev, ApplicationManagerClient appMgr)
-        {
-            this.appDev = appDev;
-            
-            this.appMgr = appMgr;
-        }
-
-        [FunctionName("SaveDAFApp")]
+        [FunctionName("SetViewType")]
         public virtual async Task<Status> Run([HttpTrigger] HttpRequest req, ILogger log,
             [SignalR(HubName = ApplicationManagementState.HUB_NAME)]IAsyncCollector<SignalRMessage> signalRMessages,
             [Blob("state-api/{headers.lcu-ent-api-key}/{headers.lcu-hub-name}/{headers.x-ms-client-principal-id}/{headers.lcu-state-key}", FileAccess.ReadWrite)] CloudBlockBlob stateBlob)
         {
-            return await stateBlob.WithStateHarness<ApplicationManagementState, SaveDAFAppRequest, ApplicationManagementStateHarness>(req, signalRMessages, log,
+            return await stateBlob.WithStateHarness<ApplicationManagementState, SetViewTypeRequest, ApplicationManagementStateHarness>(req, signalRMessages, log,
                 async (harness, reqData, actReq) =>
             {
-                log.LogInformation($"Refresh");
-
                 var stateDetails = StateUtils.LoadStateDetails(req);
 
-                await harness.SaveDAFApp(appDev, appMgr, stateDetails.EnterpriseAPIKey, reqData.DAFApp);
+                log.LogInformation($"Setting View Type: {reqData.AppType}");
+
+                await harness.SetViewType(reqData.AppType);
 
                 return Status.Success;
             });
