+++ Host/Startup.cs
@@ -14,9 +14,9 @@ using System;
 using LCU.StateAPI;
 using LCU.StateAPI.Hosting;
 
-[assembly: FunctionsStartup(typeof(LCU.State.API.NapkinIDE.ApplicationManagement.Startup))]
+[assembly: FunctionsStartup(typeof(LCU.State.API.NapkinIDE.ApplicationManagement.Host.Startup))]
 
-namespace LCU.State.API.NapkinIDE.ApplicationManagement
+namespace LCU.State.API.NapkinIDE.ApplicationManagement.Host
 {
     public class Startup : StateAPIStartup
     {
